chore: update tooling, remove clutter, streamline dev setup
