--- Makefile
@@ -1,14 +1,0 @@
-.PHONY: default
-default: run
-
-.PHONY: install
-install:
-	poetry install --develop .
-
-.PHONY: run
-run:
-	poetry run python main.py
-
-.PHONY: test
-test:
-	poetry run pytest
