+++ README.md
@@ -78,7 +78,7 @@ As soon as such a poll is closed, the results will be visible. **Beware!**: such
 The project is still actively maintained. However, since my focus shifted to other projects, I most likely won't add new features myself.  
 I'll still discuss and accept new feature requests and review feature PR's. Happy coding ;)
 
-# Acknowledgements
+## Acknowledgements
 
 <a href="https://sentry.io" ><img align="right" src="https://raw.githubusercontent.com/Nukesor/images/main/sentry.svg" alt="Packaging status"></a>
 Thanks to Sentry for providing the project with a sponsored plan (which has super generous quotas!).
@@ -139,6 +139,7 @@ Thanks to the Czech translator:
 Dependencies:
 
 - [Poetry](https://python-poetry.org/) to manage and install dependencies.
+- [Just](https://github.com/casey/just) for convience.
 - Ultimate Pollbot uses Postgres. Make sure the user has write/read rights. You can use [the provided docker-compose
  file](https://github.com/Nukesor/ultimate-poll-bot/blob/main/docker/docker-compose.yml) to set up a local development
   environment.
@@ -149,21 +150,33 @@ Dependencies:
 git clone git@github.com:nukesor/ultimate-poll-bot pollbot && cd pollbot
 ```
 
-1. Execute `poetry install` to install all dependencies.
-1. Either start the poll bot once with `poetry run python main.py` or copy the `pollbot.toml` manually to `~/.config
-/pollbot.toml` and adjust all necessary values. On Windows, the tilde (`~`) will substitute to your home directory
-, usually at `C:\Users\your.name\.config\pollbot.toml`.
-1. Run `poetry run python initdb.py` to initialize the database (or recreate it, if necessary).
-1. Start the bot by running `poetry run python main.py`.
-1. If you plan to keep your database schema up-to-date, you need to set the current alembic revision manually with
- `poetry run alembic stamp head`.
-1. Double-check if you are on the right revision `(head)` with `poetry run alembic current` and comparing it to `poetry run alembic history`.
+1. Execute `just setup` to install all dependencies.
+1. Either start the poll bot once with `just run` or copy the `pollbot.toml` manually to `~/.config/pollbot.toml`
+    and adjust all necessary values.
+    On Windows, the tilde (`~`) will substitute to your home directory, usually at `C:\Users\your.name\.config\pollbot.toml`.
+1. Run `just initdb` to initialize the database (or recreate it, if necessary) and set the migration stamp to the newest alembic head.
+1. Start the bot by running `just run`.
 
 ## Upgrading the Database
 
-If you did Step 7 in the previous section, this is how you upgrade
+This is how you upgrade:
 
 1. Stop the bot
 1. `git pull`
 1. `poetry run alembic upgrade head` to run migrations on your database
 1. Start the bot
+
+## Botfather Commands
+
+```txt
+start - Start the bot
+stop - Stop the bot
+delete_me - Remove me from the bot. Forever
+settings - Open the user settings menu
+create - Create a new poll
+cancel - Cancel poll creation
+list - List all active polls and manage them
+list_closed - List all closed polls and manage them
+notify - Activate notifications in external chats
+help - Show the help text
+```
