--- poetry.toml
@@ -1,2 +1,0 @@
-[virtualenvs]
-in-project = true
