+++ setup.cfg
@@ -2,6 +2,3 @@
 ignore = E501
 max-line-length = 120
 exclude = .git,venv,__pycache__
-
-[pycodestyle]
-max-line-length = 120
